Created separate js modules
