+++ js/modules/page-state.js
@@ -1,0 +1,3 @@
+/* eslint-disable no-unused-vars */
+const activePageState = () => {};
+const disablePageState = () => {};
